Extend the prompt for grade max feature to Moodle
Same as in D2L, when in:

- Deep linking mode
- In LTI1.3
- The feature flag is enabled

We'll prompt for the value of max points while creating assignment.
