+++ lms/product/moodle/_plugin/misc.py
@@ -1,8 +1,12 @@
-from lms.models import Assignment  # noqa: INP001
-from lms.product.plugin.misc import AssignmentConfig, MiscPlugin
+from lms.models import Assignment
+from lms.product.plugin.misc import (
+    AssignmentConfig,
+    DeepLinkingPromptForGradableMixin,
+    MiscPlugin,
+)
 
 
-class MoodleMiscPlugin(MiscPlugin):
+class MoodleMiscPlugin(DeepLinkingPromptForGradableMixin, MiscPlugin):
     def get_assignment_configuration(
         self,
         request,
