+++ tests/unit/lms/product/moodle/_plugin/misc_test.py
@@ -7,6 +7,24 @@ from tests import factories
 
 
 class TestMoodlePlugin:
+    def test_prompt_for_gradable_returns_false_for_lti_1p0(
+        self, application_instance, plugin
+    ):
+        assert plugin.deep_linking_prompt_for_gradable(application_instance) is False
+
+    @pytest.mark.parametrize("feature_flag", [True, False])
+    def test_prompt_for_gradable_returns_setting_for_lti_13(
+        self, lti_v13_application_instance, feature_flag, plugin
+    ):
+        lti_v13_application_instance.settings.set(
+            "hypothesis", "prompt_for_gradable", feature_flag
+        )
+
+        assert (
+            plugin.deep_linking_prompt_for_gradable(lti_v13_application_instance)
+            == feature_flag
+        )
+
     def test_get_assignment_configuration_outdated_db_info(
         self, plugin, pyramid_request, get_deep_linked_assignment_configuration
     ):
